+++ src/main/java/com/prgrms/bdbks/domain/star/service/DefaultStarService.java
@@ -5,7 +5,6 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.prgrms.bdbks.common.exception.EntityNotFoundException;
 import com.prgrms.bdbks.domain.star.entity.Star;
-import com.prgrms.bdbks.domain.star.mapper.StarMapper;
 import com.prgrms.bdbks.domain.star.repository.StarRepository;
 import com.prgrms.bdbks.domain.user.entity.User;
 
@@ -19,7 +18,6 @@ public class DefaultStarService implements StarService {
 	private static final short ZERO = 0;
 
 	private final StarRepository starRepository;
-	private final StarMapper starMapper;
 
 	@Override
 	@Transactional
@@ -35,10 +33,8 @@ public class DefaultStarService implements StarService {
 
 	@Override
 	public Star findByUserId(Long userId) {
-		Star star = starRepository.findByUserId(userId)
+		return starRepository.findByUserId(userId)
 			.orElseThrow(() -> new EntityNotFoundException(Star.class, userId));
-
-		return star;
 	}
 
 	@Override
@@ -49,10 +45,8 @@ public class DefaultStarService implements StarService {
 
 	@Override
 	@Transactional
-	public void updateCount(Star star, int count) {
-
-		star.updateCount(count);
-
+	public void updateCount(Long userId, int totalCount) {
+		findByUserId(userId).updateCount(totalCount);
 	}
 
 	//TODO 거래 취소, 반품 시 별은 원상복구(원래 상태로 감소)
